refactor: rename project and update package structure to 'yuklid-user-service'
