+++ pom.xml
@@ -8,11 +8,11 @@
 		<version>3.5.3</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.sudila</groupId>
-	<artifactId>microservice-user</artifactId>
+	<groupId>com.yuklid</groupId>
+	<artifactId>yuklid-user-service</artifactId>
 	<version>${revision}</version>
-	<name>microservice-user</name>
-	<description>microservice-user</description>
+	<name>yuklid-user-service</name>
+	<description>yuklid-user-service</description>
 	<url/>
 	<licenses>
 		<license/>
