--- src/main/java/com/sudila/microservice_user/auth/AuthenticationRequest.java
@@ -1,16 +1,0 @@
-package com.sudila.microservice_user.auth;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationRequest {
-
-    private String email;
-    String password;
-}
