--- src/main/java/com/sudila/microservice_user/auth/RegisterRequest.java
@@ -1,20 +1,0 @@
-package com.sudila.microservice_user.auth;
-
-import com.sudila.microservice_user.user.Role;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterRequest {
-
-    private String firstname;
-    private String lastname;
-    private String email;
-    private String password;
-    private Role role;
-}
