--- src/main/java/com/sudila/microservice_user/config/SwaggerConfig.java
@@ -1,54 +1,0 @@
-package com.sudila.microservice_user.config;
-
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.enums.SecuritySchemeIn;
-import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
-import io.swagger.v3.oas.annotations.info.Contact;
-import io.swagger.v3.oas.annotations.info.Info;
-import io.swagger.v3.oas.annotations.info.License;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import io.swagger.v3.oas.annotations.security.SecurityScheme;
-import io.swagger.v3.oas.annotations.servers.Server;
-
-@OpenAPIDefinition(
-        info = @Info(
-                contact = @Contact(
-                        name = "Sudila",
-                        email = "sudila@example.com",
-                        url = "https://github.com/your-username"
-                ),
-                description = "OpenApi documentation for Spring Security",
-                title = "OpenApi specification - Microservice User",
-                version = "1.0",
-                license = @License(
-                        name = "License name",
-                        url = "https://some-url.com"
-                ),
-                termsOfService = "Terms of service"
-        ),
-        servers = {
-                @Server(
-                        description = "Local ENV",
-                        url = "http://localhost:8080"
-                ),
-                @Server(
-                        description = "PROD ENV",
-                        url = "https://your-prod-url.com"
-                )
-        },
-        security = {
-                @SecurityRequirement(
-                        name = "bearerAuth"
-                )
-        }
-)
-@SecurityScheme(
-        name = "bearerAuth",
-        description = "JWT auth description",
-        scheme = "bearer",
-        type = SecuritySchemeType.HTTP,
-        bearerFormat = "JWT",
-        in = SecuritySchemeIn.HEADER
-)
-public class SwaggerConfig {
-}
