--- src/main/java/com/sudila/microservice_user/token/Token.java
@@ -1,36 +1,0 @@
-package com.sudila.microservice_user.token;
-
-import com.sudila.microservice_user.user.User;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-public class Token {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    public UUID id;
-
-    @Column(unique = true)
-    public String token;
-
-    @Enumerated(EnumType.STRING)
-    public TokenType tokenType = TokenType.BEARER;
-
-    public boolean revoked;
-
-    public boolean expired;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    public User user;
-}
