--- src/main/java/com/sudila/microservice_user/token/TokenRepository.java
@@ -1,20 +1,0 @@
-package com.sudila.microservice_user.token;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface TokenRepository extends JpaRepository<Token, UUID> {
-
-    @Query(value = """
-      select t from Token t inner join User u\s
-      on t.user.id = u.id\s
-      where u.id = :id and (t.expired = false or t.revoked = false)\s
-      """)
-    List<Token> findAllValidTokenByUser(UUID id);
-
-    Optional<Token> findByToken(String token);
-}
