--- src/main/java/com/sudila/microservice_user/token/TokenType.java
@@ -1,5 +1,0 @@
-package com.sudila.microservice_user.token;
-
-public enum TokenType {
-    BEARER
-}
