--- src/main/java/com/sudila/microservice_user/user/UserRepository.java
@@ -1,12 +1,0 @@
-package com.sudila.microservice_user.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface UserRepository extends JpaRepository<User, UUID> {
-
-    Optional<User> findByEmail(String email);
-
-}
