+++ src/main/java/com/yuklid/userservice/auth/AuthenticationRequest.java
@@ -1,0 +1,16 @@
+package com.yuklid.userservice.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthenticationRequest {
+
+    private String email;
+    String password;
+}
