+++ src/main/java/com/yuklid/userservice/auth/RegisterRequest.java
@@ -1,0 +1,20 @@
+package com.yuklid.userservice.auth;
+
+import com.yuklid.userservice.user.Role;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class RegisterRequest {
+
+    private String firstname;
+    private String lastname;
+    private String email;
+    private String password;
+    private Role role;
+}
