+++ src/main/java/com/yuklid/userservice/demo/DemoController.java
@@ -1,0 +1,17 @@
+package com.yuklid.userservice.demo;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/demo-controller")
+public class DemoController {
+
+    @GetMapping
+    public ResponseEntity<String> sayHello() {
+        return ResponseEntity.ok("Hello Dhanuthi from secured endpoint");
+    }
+
+}
