+++ src/main/java/com/yuklid/userservice/token/TokenType.java
@@ -1,0 +1,5 @@
+package com.yuklid.userservice.token;
+
+public enum TokenType {
+    BEARER
+}
