+++ src/main/java/com/yuklid/userservice/user/UserRepository.java
@@ -1,0 +1,12 @@
+package com.yuklid.userservice.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface UserRepository extends JpaRepository<User, UUID> {
+
+    Optional<User> findByEmail(String email);
+
+}
