+++ src/main/resources/application-ci.yml
@@ -1,6 +1,6 @@
 spring:
   application:
-    name: microservice-user
+    name: yuklid-user-service
   datasource:
     url: jdbc:postgresql://localhost:5432/user_db
     username: postgres
