--- src/test/java/com/sudila/microservice_user/MicroserviceUserApplicationTests.java
@@ -1,13 +1,0 @@
-package com.sudila.microservice_user;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MicroserviceUserApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
