+++ src/test/java/com/yuklid/userservice/YuklidUserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.yuklid.userservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class YuklidUserServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
